refactor: 멤버의 책 조회 기능 Controller 교체 #92
- BookClub으로도 검색하기에 컨트롤러를 변경하고 api 주소도 변경했습니다.
