+++ be/src/main/java/codesquad/bookkbookk/domain/book/repository/BookRepository.java
@@ -2,6 +2,7 @@ package codesquad.bookkbookk.domain.book.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
@@ -14,5 +15,6 @@ public interface BookRepository extends JpaRepository<Book, Long> {
             "ON book.id = member_book.book.id " +
             "WHERE member_book.member.id = :memberId")
     Page<Book> findBooksByMemberId(Long memberId, Pageable pageable);
+    Slice<Book> findBooksByBookClubId(Long bookClubId, Pageable pageable);
 
 }
