+++ be/src/main/java/codesquad/bookkbookk/domain/member/controller/MemberController.java
@@ -1,5 +1,6 @@
 package codesquad.bookkbookk.domain.member.controller;
 
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -7,6 +8,8 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import codesquad.bookkbookk.domain.book.data.dto.ReadBookResponse;
+import codesquad.bookkbookk.domain.book.service.BookService;
 import codesquad.bookkbookk.domain.member.data.dto.MemberResponse;
 import codesquad.bookkbookk.domain.member.data.dto.UpdateProfileRequest;
 import codesquad.bookkbookk.domain.member.data.dto.UpdateProfileResponse;
@@ -21,6 +24,7 @@ import lombok.RequiredArgsConstructor;
 public class MemberController {
 
     private final MemberService memberService;
+    private final BookService bookService;
 
     @GetMapping
     public ResponseEntity<MemberResponse> readMember(@MemberId Long memberId) {
@@ -34,6 +38,14 @@ public class MemberController {
     public ResponseEntity<UpdateProfileResponse> updateProfile(@MemberId Long memberId,
                                                                @ModelAttribute UpdateProfileRequest request) {
         UpdateProfileResponse response = memberService.updateProfile(memberId, request);
+
+        return ResponseEntity.ok()
+                .body(response);
+    }
+
+    @GetMapping("/books")
+    public ResponseEntity<ReadBookResponse> readBooks(@MemberId Long memberId, Pageable pageable) {
+        ReadBookResponse response = bookService.readBooks(memberId, pageable);
 
         return ResponseEntity.ok()
                 .body(response);
