+++ be/src/test/java/codesquad/bookkbookk/book/integration/BookTest.java
@@ -1,7 +1,5 @@
 package codesquad.bookkbookk.book.integration;
 
-import java.util.LinkedHashMap;
-
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,7 +15,6 @@ import codesquad.bookkbookk.common.jwt.JwtProvider;
 import codesquad.bookkbookk.domain.book.data.dto.CreateBookRequest;
 import codesquad.bookkbookk.domain.book.data.dto.CreateBookResponse;
 import codesquad.bookkbookk.domain.book.data.entity.Book;
-import codesquad.bookkbookk.domain.book.data.entity.MemberBook;
 import codesquad.bookkbookk.domain.book.repository.BookRepository;
 import codesquad.bookkbookk.domain.book.repository.MemberBookRepository;
 import codesquad.bookkbookk.domain.bookclub.data.entity.BookClub;
@@ -141,48 +138,6 @@ public class BookTest extends IntegrationTest {
             softAssertions.assertThat(response.statusCode()).isEqualTo(exception.getCode());
             softAssertions.assertThat(response.jsonPath().getObject("", ApiException.class).getMessage())
                     .isEqualTo(exception.getMessage());
-        });
-    }
-
-    @Test
-    @DisplayName("요청한 멤버가 참여한 책들을 페이지로 나눠서 보내준다.")
-    void readBooks() {
-        // given
-        Member member = TestDataFactory.createMember();
-        memberRepository.save(member);
-
-        BookClub bookClub = TestDataFactory.createBookClub();
-        bookClubRepository.save(bookClub);
-
-        Book book1 = TestDataFactory.createBook1(bookClub);
-        bookRepository.save(book1);
-        Book book2 = TestDataFactory.createBook2(bookClub);
-        bookRepository.save(book2);
-
-        MemberBook memberBook1 = new MemberBook(member, book1);
-        memberBookRepository.save(memberBook1);
-
-        MemberBook memberBook2 = new MemberBook(member, book2);
-        memberBookRepository.save(memberBook2);
-
-        String accessToken = jwtProvider.createAccessToken(member.getId());
-
-        // when
-        ExtractableResponse<Response> response = RestAssured
-                .given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .queryParam("page", 0)
-                .queryParam("size", 1)
-                .when()
-                .get("/api/books")
-                .then().log().all()
-                .extract();
-
-        // then
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-            softAssertions.assertThat(((LinkedHashMap) response.jsonPath().getList("books").get(0)).get("title"))
-                    .isEqualTo("신데렐라");
         });
     }
