+++ be/src/test/java/codesquad/bookkbookk/member/integration/MemberTest.java
@@ -2,6 +2,7 @@ package codesquad.bookkbookk.member.integration;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.LinkedHashMap;
 
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
@@ -15,6 +16,12 @@ import codesquad.bookkbookk.IntegrationTest;
 import codesquad.bookkbookk.common.error.exception.ApiException;
 import codesquad.bookkbookk.common.error.exception.MemberNotFoundException;
 import codesquad.bookkbookk.common.jwt.JwtProvider;
+import codesquad.bookkbookk.domain.book.data.entity.Book;
+import codesquad.bookkbookk.domain.book.data.entity.MemberBook;
+import codesquad.bookkbookk.domain.book.repository.BookRepository;
+import codesquad.bookkbookk.domain.book.repository.MemberBookRepository;
+import codesquad.bookkbookk.domain.bookclub.data.entity.BookClub;
+import codesquad.bookkbookk.domain.bookclub.repository.BookClubRepository;
 import codesquad.bookkbookk.domain.member.data.dto.MemberResponse;
 import codesquad.bookkbookk.domain.member.data.entity.Member;
 import codesquad.bookkbookk.domain.member.repository.MemberRepository;
@@ -27,6 +34,12 @@ public class MemberTest extends IntegrationTest {
 
     @Autowired
     private MemberRepository memberRepository;
+    @Autowired
+    private BookClubRepository bookClubRepository;
+    @Autowired
+    private BookRepository bookRepository;
+    @Autowired
+    private MemberBookRepository memberBookRepository;
     @Autowired
     private JwtProvider jwtProvider;
 
@@ -112,5 +125,48 @@ public class MemberTest extends IntegrationTest {
         });
 
     }
+
+    @Test
+    @DisplayName("요청한 멤버가 참여한 책들을 페이지로 나눠서 보내준다.")
+    void readBooks() {
+        // given
+        Member member = TestDataFactory.createMember();
+        memberRepository.save(member);
+
+        BookClub bookClub = TestDataFactory.createBookClub();
+        bookClubRepository.save(bookClub);
+
+        Book book1 = TestDataFactory.createBook1(bookClub);
+        bookRepository.save(book1);
+        Book book2 = TestDataFactory.createBook2(bookClub);
+        bookRepository.save(book2);
+
+        MemberBook memberBook1 = new MemberBook(member, book1);
+        memberBookRepository.save(memberBook1);
+
+        MemberBook memberBook2 = new MemberBook(member, book2);
+        memberBookRepository.save(memberBook2);
+
+        String accessToken = jwtProvider.createAccessToken(member.getId());
+
+        // when
+        ExtractableResponse<Response> response = RestAssured
+                .given().log().all()
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .queryParam("page", 0)
+                .queryParam("size", 1)
+                .when()
+                .get("/api/members/books")
+                .then().log().all()
+                .extract();
+
+        // then
+        SoftAssertions.assertSoftly(softAssertions -> {
+            softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+            softAssertions.assertThat(((LinkedHashMap) response.jsonPath().getList("books").get(0)).get("title"))
+                    .isEqualTo("신데렐라");
+        });
+    }
+
 
 }
